requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
pyjwt>=2.8.0
anthropic>=0.18.0
chromadb>=0.4.0
boto3>=1.28.0
requests>=2.31.0
psutil>=5.9.0
pydantic>=2.0.0
